[ENH] refactor param_est to live in forecasting module (#330)
* refactor param_est to forecasting

* base_classes registry

* fix examples in param_est refactor

* fix imports in examples in param_est refactor
